+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## master
 
+- Fixed all post and put methods to pass json data to Twitch API rather than form data
+- Fixed Channels.update() method to correctly pass data to Twitch
+- Fixed Channels.get() method to call 'channel' endpoint instead of 'channels'
 - Added Streams.get_streams_in_community endpoint
 - Fixed Streams.get_stream_by_user which raised exception if stream was offline. Now returns None
   if stream is offline.